完成对MediaInfoPage的改造
